@@ -5,10 +5,11 @@ description = "Storage utilities for Python 🐍"
 authors = ["radio-aktywne <it@radioaktywne.pl>"]
 
 [tool.poetry.dependencies]
-python = "^3.11"
+python = "^3.12"
 
 [tool.poetry.group.test.dependencies]
-pytest = "^8.1"
+pytest = "^8.3"
+# Asyncio support for pytest
 pytest-asyncio = "^0.23"
 
 # See: https://python-poetry.org/docs/pyproject/#poetry-and-pep-517
